requirements.txt
torch
torchvision
timm
lightning
numpy
pillow
tensorboard
